--- sql/tables/Pair_Types.sql
@@ -1,6 +1,0 @@
-CREATE TABLE Pair_Types
-(
-    id INT UNSIGNED NOT NULL UNIQUE AUTO_INCREMENT,
-    name CHAR(20) NOT NULL UNIQUE,
-    PRIMARY KEY (id)
-);
