--- sql/tables/Ticker_Types.sql
@@ -1,7 +1,0 @@
-CREATE TABLE `Ticker_Types` (
-  `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  `name` CHAR(25) NOT NULL,
-  `created_on` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `name` (`name`)
-) ENGINE=INNODB AUTO_INCREMENT=3 DEFAULT CHARSET=UTF8MB4 COLLATE=UTF8MB4_0900_AI_CI;
