+++ src/tickle/__init__.py
@@ -1,4 +1,3 @@
 from . import api as api
 from . import gui as gui
 from . import common as common
-from . import worker as worker
