+++ src/tickle/api/__init__.py
@@ -1,4 +1,3 @@
-# from .app import app as app
 """
 ********************************************************************************************
 
@@ -26,9 +25,7 @@ def _registerBlueprints(flask_app: flask.Flask):
     bp_version = flask.Blueprint('version', __name__)
 
     bp_version.register_blueprint(routes.bp_test, url_prefix='/test')
-    # bp_version.register_blueprint(routes.bp_watches, url_prefix='/watches')
     bp_version.register_blueprint(routes.bp_watches, url_prefix='/watches')
-    bp_version.register_blueprint(routes.bp_internal, url_prefix='/internal')
     bp_version.register_blueprint(routes.bp_search_tickers, url_prefix='/search/tickers')
 
     flask_app.register_blueprint(bp_version, url_prefix='/v1')
