+++ src/tickle/api/routes/__init__.py
@@ -1,5 +1,3 @@
 from .test import bp_test as bp_test
-# from .watches import bp_watches as bp_watches
-from .watches import bp_watches
-from .internal import bp_internal as bp_internal
+from .watches import bp_watches as bp_watches
 from .search_tickers import bp_search_tickers as bp_search_tickers
