+++ src/tickle/api/services/__init__.py
@@ -1,3 +1,2 @@
 from . import watches as watches
-from . import audit as audit
 from . import search_tickers as search_tickers
