--- src/tickle/api/services/tickerlib/constants.py
@@ -1,17 +1,0 @@
-"""
-********************************************************************************************
-
-Tiingo api constants/enums
-
-********************************************************************************************
-"""
-
-from __future__ import annotations
-
-class StockPriceApiUrls:
-    BASE               = 'https://api.tiingo.com'
-    IEX                = f'{BASE}/iex'
-    SEARCH_SYMBOLS     = f'{BASE}/tiingo/utilities/search'
-    CRYPTO_ALL_TICKERS = f'{BASE}/tiingo/crypto'
-    CRYPTO_PRICES      = f'{BASE}/tiingo/crypto/top'
-
