--- src/tickle/api/services/tickerlib/prices/__init__.py
@@ -1,2 +1,0 @@
-from .stocks import getStockPrices as getStockPrices
-from .crypto import getCryptoPrices as getCryptoPrices
