+++ src/tickle/api/services/watches.py
@@ -3,153 +3,132 @@
 
 Watches services.
 
-A watch is made up of:
-    a stock to keep an eye out on
-    a target price
-    a watch type (if the price falls or if it rises beyond the watch price)
-
 ********************************************************************************************
 """
 
 from __future__ import annotations
 from uuid import UUID
 import flask
-from tickle.common import serializers, utilities, BaseReturn
 from tickle.common.domain import models
-from tickle.common.domain.enums.watches import TickerTypes
-from tickle.common.domain.views.watches import ViewOpenWatchesMaps, ViewWatch, ViewWatchMap, ViewOpenWatchesList
+from tickle.common import serializers
+from tickle.common.utilities import getUUID
+from tickle.common.structs import BaseReturn
 import tickle.api.repository.watches as watches_repo
 
 #------------------------------------------------------
-# Create a new watch record from the request data
+# Create a new watch
 #------------------------------------------------------
 def createNew() -> BaseReturn:
-    watch = _getModelFromRequestForm()
-    watch.id = utilities.getUUID(False)
+    service_result = BaseReturn(successful=False)
 
-    result = BaseReturn(successful=False)
-    
-    # save the new record
+    # retrive and serialize the request's form data
     try:
-        _saveModel(watch)
+        model = _getRequestFormData()
+        model.id = getUUID(False)
     except Exception as ex:
-        result.error = ex
-        return result
+        service_result.error = ex
+        return service_result
 
-    # fetch the watch view
+    # validate the model's properties
     try:
-        view = _getView(watch.id)
-    except Exception as ex:
-        result.error = ex
-        return result
+        _validateNewModel(model)
+    except ValueError as ex:
+        service_result.error = ex
+        return service_result
 
-    result.successful = True
-    result.data = view
-    return result
+    # insert it into the database
+    try:
+        _saveWatch(model)
+    except RuntimeError as ex:
+        service_result.error = ex
+        return service_result
 
+    # all good... 
+    service_result.successful = True
+    service_result.data = model
+
+    return service_result
 
 #------------------------------------------------------
-# Create a serialized Watch model from the request's form data
+# get the request's form data 
 #------------------------------------------------------
-def _getModelFromRequestForm() -> models.Watch:
-    form = flask.request.form.to_dict()
-    return serializers.WatchSerializer(form).serialize()
+def _getRequestFormData() -> models.Watch:
+    request_data = flask.request.form.to_dict()
+    model = _serializeModel(request_data)
+
+    return model
 
 #------------------------------------------------------
-# Save the given watch model to the database
+# serialzie the given dictionary into a watch model
 #------------------------------------------------------
-def _saveModel(watch: models.Watch):
-    result = watches_repo.insert(watch)
+def _serializeModel(watch_dict: dict) -> models.Watch:
+    serializer = serializers.WatchSerializer(watch_dict)
+    return serializer.serialize()
+
+#------------------------------------------------------
+# make sure that the specified pair_id is supported with investpy
+#------------------------------------------------------
+def _validateNewModel(watch: models.Watch):
+    if _checkForNull(watch):
+        raise ValueError("Missing a required field in the request. The required fields are: 'pair_type', 'pair_id', 'price', 'watch_type', 'email'.")
+
+#------------------------------------------------------
+# Checks if any of the required Watch properties are null 
+#------------------------------------------------------
+def _checkForNull(watch: models.Watch) -> bool:
     
+    required_properties = [
+        watch.tag,
+        watch.price,
+        watch.watch_type,
+        watch.email,
+    ]
+
+    if None in required_properties:
+        return True
+    else:
+        return False
+
+#------------------------------------------------------
+# Save the watch to the database
+#------------------------------------------------------
+def _saveWatch(watch: models.Watch):
+    db_result = watches_repo.insert(watch)
+
+    if not db_result.successful:
+        raise RuntimeError(str(db_result.error))
+
+    return db_result
+
+
+#------------------------------------------------------
+# Get all open watches
+#------------------------------------------------------
+def getOpenWatches() -> list[models.Watch]:
+    db_result = watches_repo.selectAllOpen()
+
+    if not db_result.successful:
+        raise RuntimeError(str(db_result.error))
+    
+    watches = db_result.data or []
+    return watches
+
+#------------------------------------------------------
+# Close the specified watch
+#
+# Returns a bool:
+#    true: all good
+#    false: watch does not exist
+#------------------------------------------------------
+def closeWatch(watch_id: UUID) -> bool:
+    result = watches_repo.closeWatch(watch_id)
+
     if not result.successful:
         raise result.error
 
-#------------------------------------------------------
-# Get the specified WatchView from the database
-#------------------------------------------------------
-def _getView(watch_id: UUID) -> ViewWatch | None:
-    db_result = watches_repo.select(watch_id)
-
-    if not db_result.successful:
-        raise db_result.error
-
-    if not db_result.data:
-        return None
+    if result.data == 1:
+        return True
+    else:
+        return False
     
-    serializer = serializers.WatchViewSerializer(db_result.data)
-    return serializer.serialize()
 
-
-#------------------------------------------------------
-# Get a dictionary of open watches
-# The keys are the tickers 
-#------------------------------------------------------
-def getOpenWatches() -> ViewOpenWatchesMaps:
-    model_list = _getOpenWatchesModelList()
-    models_by_ticker_type = _splitModelsByType(model_list)
-    watches_dict = _toViewOpenWatchesMaps(models_by_ticker_type)
-
-    return watches_dict
-
-#------------------------------------------------------
-# Get a list of Watch models that are open
-#------------------------------------------------------
-def _getOpenWatchesModelList() -> list[ViewWatch]:
-    models = []
-    
-    for d in watches_repo.selectAllOpen().data:
-        serializer = serializers.WatchViewSerializer(d)
-        model = serializer.serialize()
-        models.append(model)
-
-    return models
-
-#------------------------------------------------------
-# Split the 1 dimensional list of models into their ticker types (stocks or crypto)
-#------------------------------------------------------
-def _splitModelsByType(open_watches: list[ViewWatch]) -> ViewOpenWatchesList:
-    stocks_list = []
-    crypto_list = []
-
-    for watch in open_watches:
-        if watch.ticker_type == TickerTypes.CRYPTO:
-            crypto_list.append(watch)
-        else:
-            stocks_list.append(watch)
-
-    result = ViewOpenWatchesList(
-        stocks = stocks_list,
-        crypto = crypto_list
-    )
-
-    return result
-
-#------------------------------------------------------
-# Transform the specified ViewOpenWatchesList into a ViewOpenWatchesMaps
-#------------------------------------------------------
-def _toViewOpenWatchesMaps(models_by_ticker_type: ViewOpenWatchesList) -> ViewOpenWatchesMaps:
-    result = ViewOpenWatchesMaps(
-        stocks = _toViewWatchMap(models_by_ticker_type.stocks),
-        crypto = _toViewWatchMap(models_by_ticker_type.crypto),
-    )
-
-    return result
-
-#------------------------------------------------------
-# Transform the given list of Watches into a dictionary of tickers as keys
-#------------------------------------------------------
-def _toViewWatchMap(models_list: list[ViewWatch]) -> ViewWatchMap:
-    watches_dict = {}
-
-    for model in models_list:
-        if model.ticker in watches_dict:
-            watches_dict[model.ticker].append(model)
-        else:
-            watches_dict.setdefault(model.ticker, [model])
-        
-    return watches_dict
-
-
-
-def closeWatch(watch_id: UUID):
-    result = watches_repo.update(watch_id)
