--- src/tickle/api/services/watches_new/__init__.py
@@ -1,4 +1,0 @@
-
-from .core import createNew as createNew
-from .core import getOpenWatches as getOpenWatches
-from .core import closeWatch
