--- src/tickle/api/services/watches_new/errors.py
@@ -1,10 +1,0 @@
-
-
-class TickleError(Exception):
-    pass
-
-
-
-
-NotFoundError = TickleError("")
-
