+++ src/tickle/common/domain/enums/watches.py
@@ -8,7 +8,6 @@ Enums for watches
 
 from __future__ import annotations
 from enum import Enum, unique
-
 from .extended_enums import ExtendedEnum
 
 @unique
@@ -20,32 +19,4 @@ class WatchTypes(ExtendedEnum):
 class TickerTypes(Enum):
     STOCKS = 1
     CRYPTO = 2
-
-
-@unique
-class PairTypes(ExtendedEnum):
-    INDICES      = 1
-    STOCKS       = 2
-    ETFS         = 3
-    FUNDS        = 4
-    COMMODITIES  = 5
-    CURRENCIES   = 6
-    CRYPTOS      = 7
-    BONDS        = 8
-    CERTIFICATES = 9
-    FXFUTURES    = 10
-
-
-@unique
-class PairTypesText(str, ExtendedEnum):
-    INDICES      = 'indices'
-    STOCKS       = 'stocks'
-    ETFS         = 'etfs'
-    FUNDS        = 'funds'
-    COMMODITIES  = 'commodities'
-    CURRENCIES   = 'currencies'
-    CRYPTOS      = 'cryptos'
-    BONDS        = 'bonds'
-    CERTIFICATES = 'certificates'
-    FXFUTURES    = 'fxfutures'
