+++ src/tickle/common/domain/models/__init__.py
@@ -1,3 +1,1 @@
-from .ticker_price import TickerPrice as TickerPrice
-from .ticker_price import TickerPriceMap as TickerPriceMap
 from .watches import Watch as Watch
