+++ src/tickle/jobs/audit/api_wrapper.py
@@ -11,9 +11,7 @@ from enum import Enum
 import requests
 from tickle.common import serializers
 from tickle.common.domain.models.watches import Watch
-from tickle.common.domain.views.watches import ViewWatch
 from tickle.common.config import configs
-from tickle.common.utilities import dumpJson
 
 API_URL_PREFIX = configs.Production.URL_API
 
@@ -49,37 +47,11 @@ def _requestOpenWatches() -> list[dict]:
     return api_response.json().get('data')
 
 #------------------------------------------------------
-# Get a list of watches that need to be closed from the API
+# Tell the api to close the specified watch
 #------------------------------------------------------
-def getWatchesToClose() -> list[ViewWatch]:
-    api_response_data = _getRequestData()
-    watches_to_close = []
-
-    for record in api_response_data:
-        serializer = serializers.WatchViewSerializer(record)
-        watches_to_close.append(serializer.serialize())
-
-    return watches_to_close
-
-
-#------------------------------------------------------
-# Get the open watches from the api
-#------------------------------------------------------
-def _getRequestData():
-    url = _getApiUrl(ApiEndpoints.AUDIT)
-    response = requests.get(url, headers=_getCustomHeaders())
-
-    if not response.ok:
-        raise Exception(str(response.text))
-
-    try:
-        data = response.json().get('data')
-    except Exception as ex:
-        print(ex)
-        data = []
-    
-    return data
-
+def closeWatch(watch_id) -> requests.Response:
+    url = f'{_getApiUrl(ApiEndpoints.WATCHES)}/{watch_id}'
+    return requests.delete(url, headers=_getCustomHeaders())
 
 #------------------------------------------------------
 # Create a custom headers dictionary
@@ -90,14 +62,6 @@ def _getCustomHeaders() -> dict:
     }
 
     return custom_headers
-
-
-#------------------------------------------------------
-# Tell the api to close the specified watch
-#------------------------------------------------------
-def closeWatch(watch_id) -> requests.Response:
-    url = f'{_getApiUrl(ApiEndpoints.WATCHES)}/{watch_id}'
-    return requests.delete(url, headers=_getCustomHeaders())
 
 
 def _getApiUrl(endpoint: ApiEndpoints) -> str:
