--- src/tickle/stockslib/test_tags.py
@@ -1,203 +1,0 @@
-
-TEST_TAGS = [
-    "/equities/us-steel-corp",
-    "/equities/tmx-group-inc-groupe-tmx-inc",
-    "/equities/xurpas-inc",
-    "/equities/united-states-steel-ar",
-    "/equities/tmx-group-inc-groupe-tmx-inc?cid=1077137",
-    "/equities/us-steel-corp?cid=993904",
-    "/equities/exxon-mobil",
-    "/equities/xcel-energy",
-    "/equities/sirius-satellite-radio-inc",
-    "/equities/xylem",
-    "/equities/xerox-corp",
-    "/equities/dentsply-intl-inc-new",
-    "/equities/xenia-hotels---resorts-inc",
-    "/equities/xpo-logistics",
-    "/equities/tessera-tech",
-    "/equities/22nd-cntry",
-    "/equities/xp-inc",
-    "/equities/xpeng-inc",
-    "/equities/xencor-inc",
-    "/equities/xeris-pharmaceuticals",
-    "/equities/qualtrics-international-inc",
-    "/equities/power-digital-infra-acquisition-ii",
-    "/equities/excelfin-acquisition",
-    "/equities/xpac-acquisition",
-    "/equities/xometry",
-    "/equities/nextgen-acquisition",
-    "/equities/china-xiangtai-food-co-ltd",
-    "/equities/xpel-inc",
-    "/equities/xenon-pharmaceuticals",
-    "/equities/exela-tech",
-    "/equities/xcel-brands-inc",
-    "/equities/xunlei-ltd-adr",
-    "/equities/vringo",
-    "/equities/xyratex-ltd",
-    "/equities/destination-xl",
-    "/equities/x-financial-a",
-    "/equities/exagen-inc",
-    "/equities/exicure",
-    "/equities/xilio-development",
-    "/equities/pivotal-investment-corporation-ii",
-    "/equities/nano-x-imaging-ltd",
-    "/equities/xponential-fitness",
-    "/equities/arsanis-inc",
-    "/equities/us-xpress-enterprises-inc",
-    "/equities/intersect-ent-inc",
-    "/equities/ait-therapeutics",
-    "/equities/xinyuan-real-estate-co-ltd",
-    "/equities/xoma-ltd",
-    "/equities/xbiotech-inc",
-    "/equities/international-nv",
-    "/equities/xtl-biophar",
-    "/equities/xenetic-biosciences,",
-    "/etfs/powershares-sp-500-ex-rate-sensiti",
-    "/etfs/deutsche-xtrackers-msci-all-china",
-    "/etfs/global-x-msci-pakistan",
-    "/etfs/recon-capital-nasdaq-100-covered",
-    "/etfs/global-x-superdividend-reit",
-    "/etfs/global-x-yieldco",
-    "/etfs/deutsche-xtrackers-csi-300-china-a",
-    "/etfs/deutsche-xtrackers-emerging-markets",
-    "/etfs/deutsche-xtrackers-high-yield-corp",
-    "/etfs/deutsche-xtrackers-investment-grade",
-    "/etfs/wisdomtree-em-ex-state-owned-ent",
-    "/etfs/deutsche-xtrackers-msci-eafe-high",
-    "/etfs/egshares-em-core-ex-china",
-    "/etfs/real-estate-select-sector-spdr",
-    "/etfs/global-x-superdividend-emerging-mar",
-    "/etfs/xtrackers-jpx-nikkei-400",
-    "/etfs/xtrackers-msci-all-world",
-    "/etfs/xtrackers-msci-all-world-high-div",
-    "/etfs/credit-suisse-x-links-gold-shares",
-    "/etfs/credit-suisse-x-links-silver-shares",
-    "/etfs/global-x-superdividend-alternatives",
-    "/equities/citigroup-capital-xiii-pref",
-    "/etfs/deutsche-xtrackers-msci-southern-eu",
-    "/etfs/recon-capital-dax-germany",
-    "/etfs/deutsche-x-trackers-msci-eurozone",
-    "/etfs/db-x-trackers-msci-europe-hdgd",
-    "/etfs/deutsche-x-trackers-russell-1000",
-    "/etfs/db-x-trackers-harvest-china",
-    "/etfs/global-x-china-energy",
-    "/etfs/global-x-china-materials",
-    "/etfs/db-x-trackers-msci-ap-x-jpn-hgd-eq",
-    "/etfs/global-x-nigeria-index",
-    "/etfs/db-x-trackers-msci-ger-hgd-eq",
-    "/etfs/horizons-sp-500-covered-call",
-    "/etfs/global-x-ftse-portugal-20",
-    "/etfs/global-x-next-emer---frontier",
-    "/etfs/global-x-superdividend-us",
-    "/etfs/global-x-mlp---energy-infrs",
-    "/etfs/db-x-trackers-muni-infras-revn",
-    "/etfs/powershares-s-p-midcap-low-vol",
-    "/etfs/pwshr-s-p-4",
-    "/etfs/deutsche-x-trackers-harvest-csi-500",
-    "/etfs/deutsche-x-trackers-ftse-developed",
-    "/etfs/global-x-cloud-computing",
-    "/etfs/global-x-msci-china-large-cap-50",
-    "/etfs/xtrackers-msci-acwi-x-us-esg-ldrseq",
-    "/etfs/amplify-international-online-retail",
-    "/etfs/xtrackers-msci-usa-esg-leaders-eq",
-    "/etfs/spdr-ssga-us-sector-rotation",
-    "/etfs/global-x-nasdaq-china-technology",
-    "/etfs/xtrackers-sp-500-esg",
-    "/etfs/global-x-msci-china-consumerstaples",
-    "/etfs/global-x-russell-2000-covered-call",
-    "/etfs/global-x-msci-china-real-estate",
-    "/etfs/hero",
-    "/etfs/global-x-adaptive-us-factor",
-    "/etfs/potx",
-    "/etfs/gnom",
-    "/etfs/chik",
-    "/etfs/bug",
-    "/equities/xai-octagon-floating-rate-alter",
-    "/etfs/bfit",
-    "/etfs/xout",
-    "/etfs/edut",
-    "/equities/credit-suisse-ag-pr",
-    "/etfs/chih",
-    "/etfs/global-x-e-commerce",
-    "/etfs/global-x-future-analytics-tech",
-    "/etfs/global-x-internet-things-them-usd",
-    "/etfs/global-x-us-infrastructure-dev",
-    "/etfs/global-x-robotics---ai-usd",
-    "/etfs/deutsche-x-track-hgh-yld-corp",
-    "/etfs/spdr-factset-innovative-technology",
-    "/etfs/spdr-sp-internet",
-    "/etfs/global-x-conscious-companies",
-    "/etfs/global-x-longevity-thematic",
-    "/etfs/global-x-msci-superdividend-eafe",
-    "/etfs/global-x-sp-500-catholic-values",
-    "/etfs/global-x-fintech-thematic",
-    "/etfs/global-x-millennials-thematic",
-    "/etfs/powershares-sp-smallcap-qual-port",
-    "/etfs/global-x-sp-500-quality-dividend",
-    "/etfs/powershares-sp-smllcap-hidiv-lowvol",
-    "/etfs/ishares-exponential-tech-mu",
-    "/etfs/global-x-us-preferred",
-    "/equities/ps-business-parks-adr-x",
-    "/etfs/xtrackers-high-beta-hi-yield-bond",
-    "/etfs/xtrackers-low-beta-hy-bond",
-    "/etfs/xtrackers-short-dur-hi-yield-bond",
-    "/etfs/global-x-auto-electric-vehicles",
-    "/etfs/communication-services-select-spdr",
-    "/etfs/silver-miners",
-    "/equities/bacterin-intl",
-    "/equities/solitario-exploration---royalty",
-    "/etfs/spdr-s-p-retail",
-    "/etfs/ishares-msci-pac-x-jp",
-    "/etfs/spdr-s-p-metals---mining",
-    "/etfs/spdr-materials-select-sector-etf",
-    "/etfs/global-x-china-consumer",
-    "/etfs/spdr-s-p-pharmaceuticals",
-    "/etfs/rydex-russell-top-50",
-    "/etfs/db-x-msci-em-currency-hedged-eq",
-    "/etfs/global-x-ftse-greece-20",
-    "/etfs/db-x-msci-japan-curr.-hedged-eq",
-    "/etfs/global-x-ftse-norway-30",
-    "/etfs/spdr-s-p-oil--gas-explor---product",
-    "/etfs/powershares-fund.-pure-small-gro.",
-    "/etfs/powersharesamental-pure-mid-core",
-    "/etfs/powersharesamental-pure-mid-growth",
-    "/etfs/powersharesamental-pure-mid-value",
-    "/etfs/powersharesamental-pure-small",
-    "/etfs/spdr-ms-technology",
-    "/etfs/global-x-interbolsa-ftcolombia-20",
-    "/etfs/spdr-s-p-aerospace---defense",
-    "/etfs/spdr-s-p-health-care-equipment",
-    "/etfs/spdr-s-p-health-care-services",
-    "/etfs/spdr-s-p-software---services",
-    "/etfs/spdr-s-p-telecom",
-    "/etfs/spdr-s-p-transportation",
-    "/etfs/spdr-select-sector---technology",
-    "/etfs/spdr-select-sector---utilities",
-    "/etfs/superincome-preferred",
-    "/etfs/global-x-social-media-index",
-    "/etfs/db-x-msci-eafe-currency-hedged-eq",
-    "/etfs/marketvectors-cef-municipal-income",
-    "/etfs/proshares-ultra-xinhua-china-25",
-    "/etfs/spdr---health-care",
-    "/etfs/spdr-s-p-oil---gas-eq---services",
-    "/etfs/spdr-s-p-semiconductor",
-    "/etfs/spdr-s-p-biotech",
-    "/etfs/spdr-s-p-homebuilders",
-    "/etfs/global-x-china-industrials",
-    "/etfs/spdr-consumer-discr.-select-sector",
-    "/etfs/spdr-energy-select-sector-fund",
-    "/etfs/financial-select-sector-spdr-fund",
-    "/etfs/industrial-sector-spdr-trust",
-    "/etfs/spdr---consumer-staples",
-    "/etfs/global-x-ftse-argentina-20",
-    "/etfs/global-x-superdividend",
-    "/etfs/global-x-copper-miners",
-    "/etfs/global-x-china-financials",
-    "/etfs/global-x-uranium",
-    "/etfs/global-x-top-guru-holdings",
-    "/etfs/global-x-mlp",
-    "/etfs/global-x-lithium",
-    "/etfs/global-x-gold-explorers",
-    "/etfs/global-x-ftse-nordic-30",
-    "/etfs/global-x-ftse-asean-40"
-]
